tests: fix windows failures after refactor (#115)
* tests: fix windows failures after refactor

Avoid MSVC stack exhaustion in large state-based test fixtures by moving those fixtures to static storage.

Skip the redirected curses render probe on Windows, where the PDCurses console backend rejects CTest file redirection, while preserving the prompt editing coverage.


* dsp: fix msvc float literal warnings


* rtltcp: fix windows regression after refactor


---------

+++ tests/protocol/dmr/test_dmr_locn_time_fallback.c
@@ -195,7 +195,7 @@ main(void) {
     long sz = ftell(f);
     fseek(f, 0, SEEK_SET);
     size_t psz = 0;
-    if (sz > 0 && (unsigned long)sz <= (unsigned long)(SIZE_MAX - 1)) {
+    if (sz > 0 && (size_t)sz <= SIZE_MAX - 1U) {
         psz = (size_t)sz;
     }
     char* buf = calloc(psz + 1u, 1u);
